move user consents test to a proper module for DIA-2940-consent-applies-fix
