+++ samples/app/src/androidTest/java/com/sourcepoint/app/v6/TestData.kt
@@ -225,85 +225,6 @@ object TestData {
         }
     """.trimIndent()
 
-    val storedConsentWithSaveAndExitChoicesAndAppliesTrueV741 = """
-        {
-            "sp.ccpa.key.consent.status": "{\n  \"applies\": true,\n  \"consentedAll\": false,\n  \"dateCreated\": \"2023-11-07T14:10:06.932Z\",\n  \"rejectedAll\": false,\n  \"rejectedCategories\": [\n    \"608bae685461ff11a2c28658\"\n  ],\n  \"rejectedVendors\": [\n    \"5dcb0978fcf9312bfc52ae91\"\n  ],\n  \"signedLspa\": false,\n  \"uspstring\": \"1YYN\",\n  \"status\": \"rejectedSome\",\n  \"uuid\": \"7014c66c-1316-4644-aa46-3d85292127a1\",\n  \"webConsentPayload\": {\n    \"actions\": [\n    ],\n    \"cookies\": [\n      {\n        \"key\": \"ccpaUUID\",\n        \"value\": \"7014c66c-1316-4644-aa46-3d85292127a1\",\n        \"setPath\": true,\n        \"maxAge\": 31536000\n      },\n      {\n        \"key\": \"ccpaReject\",\n        \"value\": true,\n        \"setPath\": true,\n        \"maxAge\": 31536000\n      },\n      {\n        \"key\": \"ccpaConsentAll\",\n        \"value\": false,\n        \"setPath\": true,\n        \"maxAge\": 31536000\n      },\n      {\n        \"key\": \"consentStatus\",\n        \"value\": \"rejectedSome\",\n        \"setPath\": true,\n        \"maxAge\": 31536000\n      }\n    ],\n    \"consentedAll\": false,\n    \"dateCreated\": \"2023-11-07T14:10:06.932Z\",\n    \"expirationDate\": \"2024-11-06T14:10:06.932Z\",\n    \"rejectedAll\": false,\n    \"rejectedCategories\": [\n      \"608bae685461ff11a2c28658\"\n    ],\n    \"rejectedVendors\": [\n      \"5dcb0978fcf9312bfc52ae91\"\n    ],\n    \"signedLspa\": false,\n    \"status\": \"rejectedSome\",\n    \"uspstring\": \"1YYN\",\n    \"uuid\": \"7014c66c-1316-4644-aa46-3d85292127a1\"\n  },\n  \"expirationDate\": \"2024-11-06T14:10:06.932Z\"\n}",
-            "IABTCF_PurposeOneTreatment": 0,
-            "sp.gdpr.key.expiration.date": "2024-11-06T14:04:38.700Z",
-            "sp.ccpa.consentUUID": "7014c66c-1316-4644-aa46-3d85292127a1",
-            "IABTCF_PublisherConsent": "0000000000",
-            "IABTCF_gdprApplies": 1,
-            "IABTCF_PublisherRestrictions4": "000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-            "IABTCF_PublisherRestrictions2": "000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-            "sp.key.messages.v7.local.state": "{\n  \"gdpr\": {\n    \"mmsCookies\": [\n      \"_sp_v1_ss=1:H4sIAAAAAAAAAItWqo5RKimOUbLKK83J0YlRSkVil4AlqmtrlXSoqiwWACMYp9h2AAAA\"\n    ],\n    \"propertyId\": 16893,\n    \"messageId\": 0\n  },\n  \"ccpa\": {\n    \"mmsCookies\": [\n      \"_sp_v1_ss=1:H4sIAAAAAAAAAItWqo5RKimOUbLKK83J0YlRSkVil4AlqmtrlWIBBrfP6SgAAAA%3D\"\n    ],\n    \"propertyId\": 16893,\n    \"messageId\": 0\n  }\n}",
-            "IABTCF_PublisherRestrictions7": "000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-            "IABTCF_VendorLegitimateInterests": "000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-            "IABTCF_PublisherCustomPurposesConsents": "0000000000",
-            "IABTCF_PolicyVersion": 4,
-            "sp.ccpa.key.expiration.date": "2024-11-06T14:04:38.746Z",
-            "IABTCF_AddtlConsent": "1~899",
-            "IABTCF_PurposeLegitimateInterests": "0100000000",
-            "IABTCF_CmpSdkID": 6,
-            "IABTCF_CmpSdkVersion": 2,
-            "IABTCF_PublisherCustomPurposesLegitimateInterests": "0000000000",
-            "sp.key.meta.data": "{\n  \"ccpa\": {\n    \"applies\": true,\n    \"sampleRate\": 1.0\n  },\n  \"gdpr\": {\n    \"additionsChangeDate\": \"2023-09-19T13:53:18.421Z\",\n    \"applies\": true,\n    \"getMessageAlways\": false,\n    \"_id\": \"608badf1a22863112f750a18\",\n    \"legalBasisChangeDate\": \"2021-06-30T15:52:45.117Z\",\n    \"version\": 47,\n    \"sampleRate\": 1.0\n  }\n}",
-            "sp.gdpr.key.sampling.result": true,
-            "sp.ccpa.key.sampling.result": true,
-            "IABUSPrivacy_String": "1YYN",
-            "IABTCF_VendorConsents": "000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-            "IABTCF_PublisherCC": "DE",
-            "IABTCF_SpecialFeaturesOptIns": "00",
-            "IABTCF_PublisherLegitimateInterests": "0000000000",
-            "IABTCF_PublisherRestrictions10": "000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-            "IABTCF_TCString": "CP03uwAP03uwAAGABBENAYEgAKCAAEAAAAYgASAAAAAAQAAAAhAAIAJA.YAAAAAAAAAAA",
-            "IABTCF_UseNonStandardTexts": 0,
-            "sp.gdpr.consentUUID": "9bf53e38-ab5d-4735-9cab-2158eaa48a8c_25",
-            "sp.key.localDataVersion": 1,
-            "IABTCF_PurposeConsents": "1010000010",
-            "sp.key.config.propertyId": 16893,
-            "sp.gdpr.key.consent.status": "{\n  \"categories\": [\n    \"608bad95d08d3112188e0e29\",\n    \"608bad95d08d3112188e0e36\",\n    \"608bad96d08d3112188e0e59\",\n    \"60b65857619abe242bed971e\"\n  ],\n  \"consentedToAll\": false,\n  \"legIntCategories\": [\n    \"608bad95d08d3112188e0e2f\"\n  ],\n  \"legIntVendors\": [\n    \"5f1b2fbeb8e05c306f2a1eb9\"\n  ],\n  \"rejectedAny\": true,\n  \"specialFeatures\": [\n  ],\n  \"vendors\": [\n    \"5f1b2fbeb8e05c306f2a1eb9\",\n    \"5ff4d000a228633ac048be41\",\n    \"5e7ced57b8e05c485246cce0\"\n  ],\n  \"addtlConsent\": \"1~899\",\n  \"consentStatus\": {\n    \"consentedAll\": false,\n    \"consentedToAny\": true,\n    \"granularStatus\": {\n      \"defaultConsent\": false,\n      \"previousOptInAll\": false,\n      \"purposeConsent\": \"SOME\",\n      \"purposeLegInt\": \"ALL\",\n      \"vendorConsent\": \"ALL\",\n      \"vendorLegInt\": \"ALL\"\n    },\n    \"hasConsentData\": true,\n    \"rejectedAny\": true,\n    \"rejectedLI\": false\n  },\n  \"dateCreated\": \"2023-11-07T14:09:57.392Z\",\n  \"euconsent\": \"CP03uwAP03uwAAGABBENAYEgAKCAAEAAAAYgASAAAAAAQAAAAhAAIAJA.YAAAAAAAAAAA\",\n  \"grants\": {\n    \"5e7ced57b8e05c485246cce0\": {\n      \"vendorGrant\": true,\n      \"purposeGrants\": {\n        \"608bad95d08d3112188e0e29\": true,\n        \"608bad95d08d3112188e0e36\": true,\n        \"608bad96d08d3112188e0e59\": true,\n        \"60b65857619abe242bed971e\": true\n      }\n    },\n    \"5f1b2fbeb8e05c306f2a1eb9\": {\n      \"vendorGrant\": true,\n      \"purposeGrants\": {\n        \"608bad95d08d3112188e0e29\": true,\n        \"608bad95d08d3112188e0e2f\": true\n      }\n    },\n    \"5ff4d000a228633ac048be41\": {\n      \"vendorGrant\": false,\n      \"purposeGrants\": {\n        \"608bad95d08d3112188e0e2f\": false,\n        \"608bad95d08d3112188e0e36\": true\n      }\n    }\n  },\n  \"TCData\": {\n    \"IABTCF_AddtlConsent\": \"1~899\",\n    \"IABTCF_CmpSdkID\": 6,\n    \"IABTCF_CmpSdkVersion\": 2,\n    \"IABTCF_PolicyVersion\": 4,\n    \"IABTCF_PublisherCC\": \"DE\",\n    \"IABTCF_PurposeOneTreatment\": 0,\n    \"IABTCF_UseNonStandardTexts\": 0,\n    \"IABTCF_TCString\": \"CP03uwAP03uwAAGABBENAYEgAKCAAEAAAAYgASAAAAAAQAAAAhAAIAJA.YAAAAAAAAAAA\",\n    \"IABTCF_VendorConsents\": \"000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\",\n    \"IABTCF_VendorLegitimateInterests\": \"000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\",\n    \"IABTCF_PurposeConsents\": \"1010000010\",\n    \"IABTCF_PurposeLegitimateInterests\": \"0100000000\",\n    \"IABTCF_SpecialFeaturesOptIns\": \"00\",\n    \"IABTCF_PublisherConsent\": \"0000000000\",\n    \"IABTCF_PublisherLegitimateInterests\": \"0000000000\",\n    \"IABTCF_PublisherCustomPurposesConsents\": \"0000000000\",\n    \"IABTCF_PublisherCustomPurposesLegitimateInterests\": \"0000000000\",\n    \"IABTCF_gdprApplies\": 1,\n    \"IABTCF_PublisherRestrictions2\": \"000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\"\n  },\n  \"uuid\": \"9bf53e38-ab5d-4735-9cab-2158eaa48a8c_25\",\n  \"webConsentPayload\": {\n    \"actions\": [\n    ],\n    \"addtlConsent\": \"1~899\",\n    \"cookies\": [\n      {\n        \"key\": \"consentUUID\",\n        \"value\": \"9bf53e38-ab5d-4735-9cab-2158eaa48a8c_25\",\n        \"maxAge\": 31536000,\n        \"shareRootDomain\": false\n      },\n      {\n        \"key\": \"consentDate\",\n        \"value\": \"2023-11-07T14:09:57.392Z\",\n        \"maxAge\": 31536000,\n        \"shareRootDomain\": false\n      }\n    ],\n    \"consentStatus\": {\n      \"rejectedAny\": true,\n      \"rejectedLI\": false,\n      \"consentedAll\": false,\n      \"granularStatus\": {\n        \"vendorConsent\": \"ALL\",\n        \"vendorLegInt\": \"ALL\",\n        \"purposeConsent\": \"SOME\",\n        \"purposeLegInt\": \"ALL\",\n        \"previousOptInAll\": false,\n        \"defaultConsent\": false\n      },\n      \"hasConsentData\": true,\n      \"consentedToAny\": true\n    },\n    \"customVendorsResponse\": {\n      \"consentedVendors\": [\n        {\n          \"_id\": \"5f1b2fbeb8e05c306f2a1eb9\",\n          \"name\": \"QuarticON\",\n          \"vendorType\": \"CUSTOM\"\n        },\n        {\n          \"_id\": \"5ff4d000a228633ac048be41\",\n          \"name\": \"Game Accounts\",\n          \"vendorType\": \"CUSTOM\"\n        }\n      ],\n      \"consentedPurposes\": [\n        {\n          \"_id\": \"608bad95d08d3112188e0e29\",\n          \"name\": \"Store and\\/or access information on a device\"\n        },\n        {\n          \"_id\": \"608bad95d08d3112188e0e36\",\n          \"name\": \"Create profiles for personalised advertising\"\n        },\n        {\n          \"_id\": \"608bad96d08d3112188e0e59\",\n          \"name\": \"Understand audiences through statistics or combinations of data from different sources\"\n        },\n        {\n          \"_id\": \"60b65857619abe242bed971e\",\n          \"name\": \"Our Custom Purpose\"\n        }\n      ],\n      \"legIntPurposes\": [\n        {\n          \"_id\": \"608bad95d08d3112188e0e2f\",\n          \"name\": \"Use limited data to select advertising\"\n        }\n      ]\n    },\n    \"dateCreated\": \"2023-11-07T14:09:57.392Z\",\n    \"expirationDate\": \"2024-11-06T14:09:57.391Z\",\n    \"euconsent\": \"CP03uwAP03uwAAGABBENAYEgAKCAAEAAAAYgASAAAAAAQAAAAhAAIAJA.YAAAAAAAAAAA\",\n    \"grants\": {\n      \"5e7ced57b8e05c485246cce0\": {\n        \"vendorGrant\": true,\n        \"purposeGrants\": {\n          \"608bad95d08d3112188e0e29\": true,\n          \"608bad95d08d3112188e0e36\": true,\n          \"608bad96d08d3112188e0e59\": true,\n          \"60b65857619abe242bed971e\": true\n        }\n      },\n      \"5f1b2fbeb8e05c306f2a1eb9\": {\n        \"vendorGrant\": true,\n        \"purposeGrants\": {\n          \"608bad95d08d3112188e0e29\": true,\n          \"608bad95d08d3112188e0e2f\": true\n        }\n      },\n      \"5ff4d000a228633ac048be41\": {\n        \"vendorGrant\": false,\n        \"purposeGrants\": {\n          \"608bad95d08d3112188e0e2f\": false,\n          \"608bad95d08d3112188e0e36\": true\n        }\n      }\n    },\n    \"uuid\": \"9bf53e38-ab5d-4735-9cab-2158eaa48a8c_25\"\n  },\n  \"expirationDate\": \"2024-11-06T14:09:57.391Z\"\n}",
-            "sp.key.messages.v7.nonKeyedLocalState": "{\n  \"ccpa\": {\n    \"_sp_v1_data\": \"585620\",\n    \"_sp_v1_p\": \"461\"\n  },\n  \"gdpr\": {\n    \"_sp_v1_data\": \"696784\",\n    \"_sp_v1_p\": \"205\"\n  }\n}"
-        }
-    """.trimIndent()
-
-    val storedConsentWithSaveAndExitChoicesAndAppliesFalseV741 = """
-        {
-            "sp.ccpa.key.consent.status": "{\n  \"applies\": false,\n  \"consentedAll\": true,\n  \"dateCreated\": \"2023-11-07T15:21:00.414Z\",\n  \"rejectedAll\": false,\n  \"rejectedCategories\": [\n  ],\n  \"rejectedVendors\": [\n  ],\n  \"signedLspa\": false,\n  \"uspstring\": \"1YNN\",\n  \"status\": \"consentedAll\",\n  \"uuid\": \"4decb53d-97ed-4350-946d-9cf3e32afa09\",\n  \"webConsentPayload\": {\n    \"actions\": [\n    ],\n    \"cookies\": [\n      {\n        \"key\": \"ccpaUUID\",\n        \"value\": \"4decb53d-97ed-4350-946d-9cf3e32afa09\",\n        \"setPath\": true,\n        \"maxAge\": 31536000\n      },\n      {\n        \"key\": \"ccpaReject\",\n        \"value\": false,\n        \"setPath\": true,\n        \"maxAge\": 31536000\n      },\n      {\n        \"key\": \"ccpaConsentAll\",\n        \"value\": true,\n        \"setPath\": true,\n        \"maxAge\": 31536000\n      },\n      {\n        \"key\": \"consentStatus\",\n        \"value\": \"consentedAll\",\n        \"setPath\": true,\n        \"maxAge\": 31536000\n      }\n    ],\n    \"consentedAll\": true,\n    \"dateCreated\": \"2023-11-07T15:21:00.414Z\",\n    \"expirationDate\": \"2024-11-06T15:21:00.414Z\",\n    \"rejectedAll\": false,\n    \"rejectedCategories\": [\n    ],\n    \"rejectedVendors\": [\n    ],\n    \"signedLspa\": false,\n    \"status\": \"consentedAll\",\n    \"uspstring\": \"1YNN\",\n    \"uuid\": \"4decb53d-97ed-4350-946d-9cf3e32afa09\"\n  },\n  \"expirationDate\": \"2024-11-06T15:21:00.414Z\"\n}",
-            "IABTCF_PurposeOneTreatment": 0,
-            "sp.ccpa.consentUUID": "4decb53d-97ed-4350-946d-9cf3e32afa09",
-            "sp.ccpa.key.message.metadata": "{\n  \"bucket\": 380,\n  \"categoryId\": 2,\n  \"messageId\": 766174,\n  \"msgDescription\": \"\",\n  \"prtnUUID\": \"0f7a409a-23de-4598-a39a-148cc3aa7325\",\n  \"subCategoryId\": 5\n}",
-            "IABTCF_PublisherConsent": "0000000000",
-            "IABTCF_gdprApplies": 0,
-            "sp.key.messages.v7.local.state": "{\n  \"gdpr\": {\n    \"mmsCookies\": [\n      \"_sp_v1_ss=1:H4sIAAAAAAAAAItWqo5RKimOUbKKBjLyQAyD2lidGKVUEDOvNCcHyC4BK6iurVWKBQAW54XRMAAAAA%3D%3D\"\n    ],\n    \"propertyId\": 16893,\n    \"messageId\": 902943\n  },\n  \"ccpa\": {\n    \"mmsCookies\": [\n      \"_sp_v1_ss=1:H4sIAAAAAAAAAItWqo5RKimOUbKKBjLyQAyD2lidGKVUEDOvNCcHyC4BK6iurVWKBQAW54XRMAAAAA%3D%3D\"\n    ],\n    \"propertyId\": 16893,\n    \"messageId\": 766174\n  }\n}",
-            "IABTCF_VendorLegitimateInterests": "000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-            "IABTCF_PublisherCustomPurposesConsents": "0000000000",
-            "IABTCF_PolicyVersion": 4,
-            "sp.ccpa.key.expiration.date": "2024-11-06T15:17:21.767Z",
-            "IABTCF_AddtlConsent": "1~",
-            "IABTCF_PurposeLegitimateInterests": "0000000000",
-            "IABTCF_CmpSdkID": 6,
-            "IABTCF_CmpSdkVersion": 2,
-            "IABTCF_PublisherCustomPurposesLegitimateInterests": "0000000000",
-            "sp.key.meta.data": "{\n  \"ccpa\": {\n    \"applies\": false,\n    \"sampleRate\": 1.0\n  },\n  \"gdpr\": {\n    \"additionsChangeDate\": \"2023-09-19T13:53:18.421Z\",\n    \"applies\": false,\n    \"getMessageAlways\": false,\n    \"_id\": \"608badf1a22863112f750a18\",\n    \"legalBasisChangeDate\": \"2021-06-30T15:52:45.117Z\",\n    \"version\": 47,\n    \"sampleRate\": 1.0\n  }\n}",
-            "sp.gdpr.key.sampling.result": true,
-            "IABUSPrivacy_String": "1YNN",
-            "sp.ccpa.key.sampling.result": true,
-            "IABTCF_VendorConsents": "000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-            "IABTCF_PublisherCC": "UA",
-            "IABTCF_SpecialFeaturesOptIns": "00",
-            "IABTCF_PublisherLegitimateInterests": "0000000000",
-            "IABTCF_TCString": "",
-            "sp.gdpr.consentUUID": "d5411f32-2edd-46a8-8359-f95dadc82346_25",
-            "IABTCF_UseNonStandardTexts": 0,
-            "sp.key.localDataVersion": 1,
-            "sp.gdpr.key.message.metadata": "{\n  \"bucket\": 126,\n  \"categoryId\": 1,\n  \"messageId\": 902943,\n  \"msgDescription\": \"\",\n  \"prtnUUID\": \"20b4895f-1d3f-438e-bb8f-912e9ccef018\",\n  \"subCategoryId\": 5\n}",
-            "IABTCF_PurposeConsents": "0000000000",
-            "sp.key.config.propertyId": 16893,
-            "sp.gdpr.key.consent.status": "{\n  \"categories\": [\n  ],\n  \"consentedToAll\": false,\n  \"legIntCategories\": [\n  ],\n  \"legIntVendors\": [\n  ],\n  \"rejectedAny\": true,\n  \"specialFeatures\": [\n  ],\n  \"vendors\": [\n  ],\n  \"addtlConsent\": \"1~\",\n  \"consentStatus\": {\n    \"consentedAll\": false,\n    \"consentedToAny\": true,\n    \"granularStatus\": {\n      \"defaultConsent\": false,\n      \"previousOptInAll\": false,\n      \"purposeConsent\": \"NONE\",\n      \"purposeLegInt\": \"NONE\",\n      \"vendorConsent\": \"NONE\",\n      \"vendorLegInt\": \"NONE\"\n    },\n    \"hasConsentData\": true,\n    \"rejectedAny\": true,\n    \"rejectedLI\": true\n  },\n  \"dateCreated\": \"2023-11-07T15:20:56.458Z\",\n  \"euconsent\": \"CP03uwAP03uwAAGABBENAYEgAAAAAAAAACgAAAAAAAAA.YAAAAAAAAAAA\",\n  \"grants\": {\n    \"5e7ced57b8e05c485246cce0\": {\n      \"vendorGrant\": false,\n      \"purposeGrants\": {\n        \"608bad95d08d3112188e0e29\": false,\n        \"608bad95d08d3112188e0e36\": false,\n        \"608bad96d08d3112188e0e59\": false,\n        \"60b65857619abe242bed971e\": false\n      }\n    },\n    \"5f1b2fbeb8e05c306f2a1eb9\": {\n      \"vendorGrant\": false,\n      \"purposeGrants\": {\n        \"608bad95d08d3112188e0e29\": false,\n        \"608bad95d08d3112188e0e2f\": false\n      }\n    },\n    \"5ff4d000a228633ac048be41\": {\n      \"vendorGrant\": false,\n      \"purposeGrants\": {\n        \"608bad95d08d3112188e0e2f\": false,\n        \"608bad95d08d3112188e0e36\": false\n      }\n    }\n  },\n  \"TCData\": {\n    \"IABTCF_AddtlConsent\": \"1~\",\n    \"IABTCF_CmpSdkID\": 6,\n    \"IABTCF_CmpSdkVersion\": 2,\n    \"IABTCF_PolicyVersion\": 4,\n    \"IABTCF_PublisherCC\": \"UA\",\n    \"IABTCF_PurposeOneTreatment\": 0,\n    \"IABTCF_UseNonStandardTexts\": 0,\n    \"IABTCF_TCString\": \"\",\n    \"IABTCF_VendorConsents\": \"000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\",\n    \"IABTCF_VendorLegitimateInterests\": \"000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\",\n    \"IABTCF_PurposeConsents\": \"0000000000\",\n    \"IABTCF_PurposeLegitimateInterests\": \"0000000000\",\n    \"IABTCF_SpecialFeaturesOptIns\": \"00\",\n    \"IABTCF_PublisherConsent\": \"0000000000\",\n    \"IABTCF_PublisherLegitimateInterests\": \"0000000000\",\n    \"IABTCF_PublisherCustomPurposesConsents\": \"0000000000\",\n    \"IABTCF_PublisherCustomPurposesLegitimateInterests\": \"0000000000\",\n    \"IABTCF_gdprApplies\": 0\n  },\n  \"uuid\": \"d5411f32-2edd-46a8-8359-f95dadc82346_25\",\n  \"webConsentPayload\": {\n    \"actions\": [\n    ],\n    \"addtlConsent\": \"1~\",\n    \"cookies\": [\n      {\n        \"key\": \"consentUUID\",\n        \"value\": \"d5411f32-2edd-46a8-8359-f95dadc82346_25\",\n        \"maxAge\": 31536000,\n        \"shareRootDomain\": false\n      },\n      {\n        \"key\": \"consentDate\",\n        \"value\": \"2023-11-07T15:20:56.458Z\",\n        \"maxAge\": 31536000,\n        \"shareRootDomain\": false\n      }\n    ],\n    \"consentStatus\": {\n      \"rejectedAny\": true,\n      \"rejectedLI\": true,\n      \"consentedAll\": false,\n      \"granularStatus\": {\n        \"vendorConsent\": \"NONE\",\n        \"vendorLegInt\": \"NONE\",\n        \"purposeConsent\": \"NONE\",\n        \"purposeLegInt\": \"NONE\",\n        \"previousOptInAll\": false,\n        \"defaultConsent\": false\n      },\n      \"hasConsentData\": true,\n      \"consentedToAny\": true\n    },\n    \"customVendorsResponse\": {\n      \"consentedVendors\": [\n      ],\n      \"consentedPurposes\": [\n      ],\n      \"legIntPurposes\": [\n      ]\n    },\n    \"dateCreated\": \"2023-11-07T15:20:56.458Z\",\n    \"expirationDate\": \"2024-11-06T15:20:56.457Z\",\n    \"euconsent\": \"CP03uwAP03uwAAGABBENAYEgAAAAAAAAACgAAAAAAAAA.YAAAAAAAAAAA\",\n    \"grants\": {\n      \"5e7ced57b8e05c485246cce0\": {\n        \"vendorGrant\": false,\n        \"purposeGrants\": {\n          \"608bad95d08d3112188e0e29\": false,\n          \"608bad95d08d3112188e0e36\": false,\n          \"608bad96d08d3112188e0e59\": false,\n          \"60b65857619abe242bed971e\": false\n        }\n      },\n      \"5f1b2fbeb8e05c306f2a1eb9\": {\n        \"vendorGrant\": false,\n        \"purposeGrants\": {\n          \"608bad95d08d3112188e0e29\": false,\n          \"608bad95d08d3112188e0e2f\": false\n        }\n      },\n      \"5ff4d000a228633ac048be41\": {\n        \"vendorGrant\": false,\n        \"purposeGrants\": {\n          \"608bad95d08d3112188e0e2f\": false,\n          \"608bad95d08d3112188e0e36\": false\n        }\n      }\n    },\n    \"uuid\": \"d5411f32-2edd-46a8-8359-f95dadc82346_25\"\n  },\n  \"expirationDate\": \"2024-11-06T15:20:56.457Z\"\n}",
-            "sp.key.messages.v7.nonKeyedLocalState": "{\n  \"ccpa\": {\n    \"_sp_v1_data\": \"585620\",\n    \"_sp_v1_p\": \"380\"\n  },\n  \"gdpr\": {\n    \"_sp_v1_data\": \"696784\",\n    \"_sp_v1_p\": \"126\"\n  }\n}"
-        }
-    """.trimIndent()
-
     val storedConsentWithAuthIdAndPropertyIdV741 = """
         {
             "sp.gdpr.authId": "ee7ea3b8-9609-4ba4-be07-0986d32cdd1e",
